requirements.txt
pulsectl~=20.5.1
python-xlib~=0.27
pygobject~=3.36.1